Rename data2 to data
