+++ app/src/main/java/co/touchlab/droidconandroid/shared/data/AppPrefs.java
@@ -1,4 +1,4 @@
-package co.touchlab.droidconandroid.shared.data2;
+package co.touchlab.droidconandroid.shared.data;
 
 import android.content.Context;
 import android.content.SharedPreferences;
