+++ app/src/main/java/co/touchlab/droidconandroid/shared/data/DroidconDatabase.java
@@ -1,12 +1,12 @@
-package co.touchlab.droidconandroid.shared.data2;
+package co.touchlab.droidconandroid.shared.data;
 
 import android.arch.persistence.room.Database;
 import android.arch.persistence.room.RoomDatabase;
 
-import co.touchlab.droidconandroid.shared.data2.dao.BlockDao;
-import co.touchlab.droidconandroid.shared.data2.dao.EventDao;
-import co.touchlab.droidconandroid.shared.data2.dao.EventSpeakerDao;
-import co.touchlab.droidconandroid.shared.data2.dao.UserAccountDao;
+import co.touchlab.droidconandroid.shared.data.dao.BlockDao;
+import co.touchlab.droidconandroid.shared.data.dao.EventDao;
+import co.touchlab.droidconandroid.shared.data.dao.EventSpeakerDao;
+import co.touchlab.droidconandroid.shared.data.dao.UserAccountDao;
 
 @Database(entities = { Block.class, Event.class, EventSpeaker.class, UserAccount.class }, version = DroidconDatabase.DATABASE_VERSION)
 public abstract class DroidconDatabase extends RoomDatabase {
