+++ app/src/main/java/co/touchlab/droidconandroid/shared/data/dao/EventSpeakerDao.java
@@ -1,4 +1,4 @@
-package co.touchlab.droidconandroid.shared.data2.dao;
+package co.touchlab.droidconandroid.shared.data.dao;
 
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Insert;
@@ -7,7 +7,7 @@ import android.arch.persistence.room.Query;
 
 import java.util.List;
 
-import co.touchlab.droidconandroid.shared.data2.EventSpeaker;
+import co.touchlab.droidconandroid.shared.data.EventSpeaker;
 
 @Dao
 public interface EventSpeakerDao {
