+++ app/src/main/java/co/touchlab/droidconandroid/shared/data/dao/UserAccountDao.java
@@ -1,4 +1,4 @@
-package co.touchlab.droidconandroid.shared.data2.dao;
+package co.touchlab.droidconandroid.shared.data.dao;
 
 import android.arch.persistence.room.Dao;
 import android.arch.persistence.room.Insert;
@@ -7,7 +7,7 @@ import android.arch.persistence.room.Query;
 
 import java.util.List;
 
-import co.touchlab.droidconandroid.shared.data2.UserAccount;
+import co.touchlab.droidconandroid.shared.data.UserAccount;
 
 @Dao
 public interface UserAccountDao {
