+++ app/src/main/java/co/touchlab/droidconandroid/shared/network/dao/NetworkBlock.java
@@ -1,6 +1,6 @@
 package co.touchlab.droidconandroid.shared.network.dao;
 
-import co.touchlab.droidconandroid.shared.data2.Block;
+import co.touchlab.droidconandroid.shared.data.Block;
 
 /**
  * Created by izzyoji :) on 8/12/15.
