+++ app/src/main/java/co/touchlab/droidconandroid/shared/network/dao/NetworkEvent.java
@@ -4,7 +4,7 @@ import android.arch.persistence.room.Ignore;
 
 import java.util.List;
 
-import co.touchlab.droidconandroid.shared.data2.Event;
+import co.touchlab.droidconandroid.shared.data.Event;
 
 /**
  * Created by kgalligan on 7/19/14.
