+++ app/src/main/java/co/touchlab/droidconandroid/shared/presenter/UserDetailHost.java
@@ -1,6 +1,6 @@
 package co.touchlab.droidconandroid.shared.presenter;
 
-import co.touchlab.droidconandroid.shared.data2.UserAccount;
+import co.touchlab.droidconandroid.shared.data.UserAccount;
 import io.reactivex.annotations.NonNull;
 
 public interface UserDetailHost {
