+++ app/src/main/java/co/touchlab/droidconandroid/shared/tasks/persisted/RefreshScheduleJob.java
@@ -8,8 +8,8 @@ import com.birbit.android.jobqueue.Params;
 import com.birbit.android.jobqueue.RetryConstraint;
 
 import co.touchlab.droidconandroid.CrashReport;
-import co.touchlab.droidconandroid.shared.data2.AppPrefs;
-import co.touchlab.droidconandroid.shared.data2.DatabaseHelper;
+import co.touchlab.droidconandroid.shared.data.AppPrefs;
+import co.touchlab.droidconandroid.shared.data.DatabaseHelper;
 import co.touchlab.droidconandroid.shared.network.DataHelper;
 import co.touchlab.droidconandroid.shared.network.RefreshScheduleDataRequest;
 import co.touchlab.droidconandroid.shared.network.dao.Convention;
