+++ app/src/main/kotlin/co/touchlab/droidconandroid/EventDetailFragment.kt
@@ -14,7 +14,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import co.touchlab.droidconandroid.shared.data2.*
+import co.touchlab.droidconandroid.shared.data.*
 import co.touchlab.droidconandroid.shared.interactors.EventDetailInteractor
 import co.touchlab.droidconandroid.shared.interactors.RsvpInteractor
 import co.touchlab.droidconandroid.shared.interactors.UpdateAlertsInteractor
