+++ app/src/main/kotlin/co/touchlab/droidconandroid/ScheduleDataFragment.kt
@@ -9,9 +9,9 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.touchlab.droidconandroid.shared.data2.AppPrefs
-import co.touchlab.droidconandroid.shared.data2.Event
-import co.touchlab.droidconandroid.shared.data2.DatabaseHelper
+import co.touchlab.droidconandroid.shared.data.AppPrefs
+import co.touchlab.droidconandroid.shared.data.Event
+import co.touchlab.droidconandroid.shared.data.DatabaseHelper
 import co.touchlab.droidconandroid.shared.interactors.UpdateAlertsInteractor
 import co.touchlab.droidconandroid.shared.presenter.*
 import co.touchlab.droidconandroid.ui.EventAdapter
