+++ app/src/main/kotlin/co/touchlab/droidconandroid/UserDetailFragment.kt
@@ -15,8 +15,8 @@ import android.text.Html
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import co.touchlab.droidconandroid.shared.data2.DatabaseHelper
-import co.touchlab.droidconandroid.shared.data2.UserAccount
+import co.touchlab.droidconandroid.shared.data.DatabaseHelper
+import co.touchlab.droidconandroid.shared.data.UserAccount
 import co.touchlab.droidconandroid.shared.interactors.FindUserInteractor
 import co.touchlab.droidconandroid.shared.network.DataHelper
 import co.touchlab.droidconandroid.shared.network.FindUserRequest
