+++ app/src/main/kotlin/co/touchlab/droidconandroid/ui/EventAdapter.kt
@@ -9,6 +9,8 @@ import android.view.ViewGroup
 import co.touchlab.droidconandroid.R
 import co.touchlab.droidconandroid.isOdd
 import co.touchlab.droidconandroid.setViewVisibility
+import co.touchlab.droidconandroid.shared.data.Block
+import co.touchlab.droidconandroid.shared.data.Event
 import co.touchlab.droidconandroid.shared.presenter.HourBlock
 import co.touchlab.droidconandroid.shared.utils.EventBusExt
 import co.touchlab.droidconandroid.shared.utils.EventUtils
@@ -64,7 +66,7 @@ class EventAdapter(private val context: Context,
                 EventUtils.styleEventRow(scheduleBlockHour, dataSet, holder, allEvents)
 
                 if (!scheduleBlockHour.timeBlock.isBlock) {
-                    holder.setOnClickListener { eventClickListener.onEventClick(scheduleBlockHour.timeBlock as co.touchlab.droidconandroid.shared.data2.Event) }
+                    holder.setOnClickListener { eventClickListener.onEventClick(scheduleBlockHour.timeBlock as Event) }
                 }
             }
         }
@@ -82,8 +84,8 @@ class EventAdapter(private val context: Context,
 
         val item = filteredData[adjustedPosition]?.timeBlock
         when (item) {
-            is co.touchlab.droidconandroid.shared.data2.Event -> return if (item.isPast) VIEW_TYPE_PAST_EVENT else VIEW_TYPE_EVENT
-            is co.touchlab.droidconandroid.shared.data2.Block -> return VIEW_TYPE_BLOCK
+            is Event -> return if (item.isPast) VIEW_TYPE_PAST_EVENT else VIEW_TYPE_EVENT
+            is Block -> return VIEW_TYPE_BLOCK
             else -> throw UnsupportedOperationException()
         }
     }
@@ -195,7 +197,7 @@ class EventAdapter(private val context: Context,
 }
 
 interface EventClickListener {
-    fun onEventClick(event: co.touchlab.droidconandroid.shared.data2.Event)
+    fun onEventClick(event: Event)
 }
 
 data class UpdateAllowNotificationEvent(val allow: Boolean)
